Updated contestSync to ensure synchronization

+++ src/main/java/com/algolens/algo_lens/dtos/contest/CodeforcesContestItemDTO.java
@@ -19,6 +19,6 @@ public class CodeforcesContestItemDTO implements Serializable {
     private String phase;
     private boolean frozen;
     private int durationSeconds;
-    private int startTimeSeconds;
+    private long startTimeSeconds;
     private int relativeTimeSeconds;
 }
